Ido rcheck

+++ backend/routes/posts.routes.js
@@ -1,4 +1,5 @@
 const router = require('express').Router();
+const session = require('express-session');
 const { Post } = require('../db/models');
 
 router.get('/', async (req, res) => {
@@ -13,13 +14,15 @@ router.get('/', async (req, res) => {
 router.post('/', async (req, res) => {
   try {
     const { img, title, description } = req.body;
-    const Posts = await Post.create({
-      img,
-      title,
-      description,
-      userId: req.session.userId,
-    });
-    res.json(Posts);
+    if (img !== '' && title !== '' && description !== '') {
+      const Posts = await Post.create({
+        img,
+        title,
+        description,
+        userId: req.session.userId,
+      });
+      res.json(Posts);
+    }
   } catch ({ message }) {
     res.status(500).json(message);
   }
@@ -28,8 +31,13 @@ router.post('/', async (req, res) => {
 router.delete('/:id', async (req, res) => {
   try {
     const { id } = req.params;
-    const result = await Post.destroy({ where: { id } });
-    res.json(Number(id));
+    const idor = await Post.findOne({
+      where: { userId: req.session.userId },
+    });
+    if (req.session.userId === 1 || req.session.userId === idor.userId) {
+      const result = await Post.destroy({ where: { id } });
+      res.json(Number(id));
+    }
   } catch ({ message }) {
     res.status(500).json(message);
   }
@@ -39,9 +47,17 @@ router.put('/:id', async (req, res) => {
   try {
     const { id } = req.params;
     const { description, img, title } = req.body;
-    await Post.update({ description, img, title }, { where: { id } });
-    const data = await Post.findOne({ where: { id } });
-    res.json(data);
+
+    if (img !== '' && title !== '' && description !== '') {
+      const idor = await Post.findOne({
+        where: { userId: req.session.userId },
+      });
+      if (req.session.userId === 1 || req.session.userId === idor.userId) {
+        await Post.update({ description, img, title }, { where: { id } });
+        const data = await Post.findOne({ where: { id } });
+        res.json(data);
+      }
+    }
   } catch (message) {
     res.status(500).json({ message: 'Crushed' });
   }
